Refactor: Split large JavaScript file into modules
- Separated concerns into ui.js, imageFilters.js, videoGenerator.js, and utils.js.
- Updated index.html to use js/main.js as the entry point (type="module").
- Ensured proper import/export structure for ES6 modules.
- Preserved all existing functionality including image loading, text overlay, image filtering, effect sequencing, and video generation with MediaRecorder.
- Improved code organization, maintainability, and testability.
