+++ index.html
@@ -145,6 +145,6 @@
         <div id="onPageConsole"></div>
     </div>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/main.js"></script>
 </body>
 </html>
